TelegramClone/requirements.txt
flask
flask-socketio
flask-sqlalchemy
eventlet
kivy
python-socketio
requests
